fix: 상담톡 스케쥴 유형 삭제 시 예외처리
주간/일별 스케쥴러에 등록된 유형은 삭제할 수 없게 수정
스케쥴러를 불러오는 로직이 innerJoin으로 구성되어 있어서 유형이 삭제된 경우, DB에는 존재하나 화면단에서는 불러오지 않게 됨

+++ api-server/src/main/java/kr/co/eicn/ippbx/server/controller/api/v1/admin/wtalk/schedule/WtalkScheduleGroupApiController.java
@@ -11,6 +11,7 @@ import kr.co.eicn.ippbx.server.controller.api.ApiBaseController;
 import kr.co.eicn.ippbx.server.repository.eicn.WtalkMentRepository;
 import kr.co.eicn.ippbx.server.repository.eicn.WtalkScheduleGroupListRepository;
 import kr.co.eicn.ippbx.server.repository.eicn.WtalkScheduleGroupRepository;
+import kr.co.eicn.ippbx.server.repository.eicn.WtalkScheduleInfoRepository;
 import kr.co.eicn.ippbx.util.JsonResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -40,6 +41,7 @@ public class WtalkScheduleGroupApiController extends ApiBaseController {
 	private final WtalkScheduleGroupRepository repository;
 	private final WtalkScheduleGroupListRepository wtalkScheduleGroupListRepository;
 	private final WtalkMentRepository wtalkMentRepository;
+    private final WtalkScheduleInfoRepository wtalkScheduleInfoRepository;
 
 	/**
 	 * 스케쥴유형 목록조회
@@ -66,6 +68,9 @@ public class WtalkScheduleGroupApiController extends ApiBaseController {
 	 */
 	@DeleteMapping("{parent}")
 	public ResponseEntity<JsonResult<Void>> delete(@PathVariable Integer parent) {
+		if (wtalkScheduleInfoRepository.isExistGroupId(parent))
+			throw new IllegalStateException("주간/일별 스케쥴러에 등록된 유형은 삭제할 수 없습니다.");
+
 		repository.deleteOnIfNullThrow(parent);
 		return ResponseEntity.ok(create());
 	}
